- refactor element_change -> atom_changes - add readme text.
