+++ src/kartograf/filters/__init__.py
@@ -1,4 +1,4 @@
-from .element_change import (
+from .atom_changes import (
     filter_atoms_h_only_h_mapped,
     filter_element_changes,
     filter_hybridization_changes,
